refactor: extract auth components to src/components/auth/
Move AuthModal, SavePoemButton, DownvoteButton, and SavedPoemsView
(~305 lines total) to src/components/auth/.

Remove 10 unused lucide-react icon imports from app.jsx (Heart, LogIn,
User, ThumbsDown, Compass, ALargeSmall, LayoutGrid, Quote, Globe,
BookOpen).

app.jsx: 4097 → 3778 lines (-2361 total from original 6139, 38% reduction)
All 316 tests pass. Zero behavior change.
