+++ src/app.jsx
@@ -2,23 +2,13 @@ import { useState, useEffect, useRef, useMemo } from 'react';
 import {
   Play,
   Pause,
-  BookOpen,
   Volume2,
   ChevronDown,
-  Quote,
-  Globe,
   Loader2,
   Search,
   X,
-  LayoutGrid,
   Sparkles,
   Feather,
-  Compass,
-  Heart,
-  LogIn,
-  User,
-  ThumbsDown,
-  ALargeSmall,
   ScrollText,
   Shuffle,
   Paintbrush,
@@ -47,6 +37,10 @@ import ShortcutHelp from './components/ShortcutHelp.jsx';
 import SplashScreen from './components/SplashScreen.jsx';
 import InsightsDrawer from './components/InsightsDrawer.jsx';
 import VerticalSidebar from './components/VerticalSidebar.jsx';
+import AuthModal from './components/auth/AuthModal.jsx';
+import SavePoemButton from './components/auth/SavePoemButton.jsx';
+import DownvoteButton from './components/auth/DownvoteButton.jsx';
+import SavedPoemsView from './components/auth/SavedPoemsView.jsx';
 
 const apiUrl = import.meta.env.VITE_API_URL || 'http://localhost:3001';
 
@@ -731,321 +725,8 @@ const prefetchManager = {
 
 // SplashScreen imported from ./components/SplashScreen.jsx
 // (SplashScreen body removed — component now in separate file)
-/* =============================================================================
-  AUTH COMPONENTS
-  =============================================================================
-*/
 
-const AuthModal = ({ isOpen, onClose, onSignInWithGoogle, theme, message }) => {
-  if (!isOpen) return null;
-
-  return (
-    <div
-      className="fixed inset-0 z-50 flex items-center justify-center p-4"
-      style={{
-        background:
-          'radial-gradient(ellipse at center, rgba(197,160,89,0.08) 0%, rgba(0,0,0,0.7) 100%)',
-      }}
-      onClick={onClose}
-    >
-      <div
-        className="relative w-full max-w-sm rounded-2xl overflow-hidden shadow-2xl"
-        style={{
-          background: 'linear-gradient(145deg, rgba(20,18,15,0.97) 0%, rgba(12,12,14,0.98) 100%)',
-          border: '1px solid rgba(197,160,89,0.25)',
-        }}
-        onClick={(e) => e.stopPropagation()}
-      >
-        {/* Gold accent line */}
-        <div
-          style={{
-            height: '2px',
-            background:
-              'linear-gradient(90deg, transparent, rgba(197,160,89,0.6), rgba(212,180,120,0.8), rgba(197,160,89,0.6), transparent)',
-          }}
-        />
-
-        <div className="px-8 pt-8 pb-6">
-          <button
-            onClick={onClose}
-            className="absolute top-4 right-4 p-2 rounded-full hover:bg-white/10 transition-colors"
-            aria-label="Close"
-          >
-            <X size={18} style={{ color: 'rgba(197,160,89,0.6)' }} />
-          </button>
-
-          {/* Greeting */}
-          <div className="text-center mb-6">
-            <h2 className="font-amiri text-3xl mb-2" style={{ color: '#C5A059' }}>
-              مرحباً
-            </h2>
-            <p className="font-brand-en text-sm text-stone-400 leading-relaxed">
-              {message || 'Sign in to save poems and preferences'}
-            </p>
-          </div>
-
-          {/* Decorative divider */}
-          <div className="flex items-center gap-3 mb-6">
-            <div className="flex-1 h-px bg-gradient-to-r from-transparent to-[#C5A059]/30" />
-            <Feather size={12} style={{ color: 'rgba(197,160,89,0.4)' }} />
-            <div className="flex-1 h-px bg-gradient-to-l from-transparent to-[#C5A059]/30" />
-          </div>
-
-          {/* Google sign-in button */}
-          <button
-            onClick={onSignInWithGoogle}
-            className="w-full py-3.5 px-5 rounded-xl transition-all duration-200 flex items-center justify-center gap-3 hover:scale-[1.02] active:scale-[0.98]"
-            style={{
-              background:
-                'linear-gradient(135deg, rgba(197,160,89,0.12) 0%, rgba(197,160,89,0.06) 100%)',
-              border: '1px solid rgba(197,160,89,0.3)',
-              color: '#D4C8B0',
-              fontFamily: "'Inter', 'Segoe UI', system-ui, -apple-system, sans-serif",
-              fontSize: '14px',
-              fontWeight: 500,
-              letterSpacing: '0.01em',
-            }}
-          >
-            <svg className="w-5 h-5" viewBox="0 0 24 24">
-              <path
-                fill="#4285F4"
-                d="M22.56 12.25c0-.78-.07-1.53-.2-2.25H12v4.26h5.92c-.26 1.37-1.04 2.53-2.21 3.31v2.77h3.57c2.08-1.92 3.28-4.74 3.28-8.09z"
-              />
-              <path
-                fill="#34A853"
-                d="M12 23c2.97 0 5.46-.98 7.28-2.66l-3.57-2.77c-.98.66-2.23 1.06-3.71 1.06-2.86 0-5.29-1.93-6.16-4.53H2.18v2.84C3.99 20.53 7.7 23 12 23z"
-              />
-              <path
-                fill="#FBBC05"
-                d="M5.84 14.09c-.22-.66-.35-1.36-.35-2.09s.13-1.43.35-2.09V7.07H2.18C1.43 8.55 1 10.22 1 12s.43 3.45 1.18 4.93l2.85-2.22.81-.62z"
-              />
-              <path
-                fill="#EA4335"
-                d="M12 5.38c1.62 0 3.06.56 4.21 1.64l3.15-3.15C17.45 2.09 14.97 1 12 1 7.7 1 3.99 3.47 2.18 7.07l3.66 2.84c.87-2.6 3.3-4.53 6.16-4.53z"
-              />
-            </svg>
-            Continue with Google
-          </button>
-
-          <p
-            className="mt-5 text-center text-[10px] text-stone-600"
-            style={{ fontFamily: "'Inter', 'Segoe UI', system-ui, -apple-system, sans-serif" }}
-          >
-            By signing in, you agree to our Terms of Service
-          </p>
-        </div>
-
-        {/* Gold accent bottom */}
-        <div
-          style={{
-            height: '1px',
-            background: 'linear-gradient(90deg, transparent, rgba(197,160,89,0.3), transparent)',
-          }}
-        />
-      </div>
-    </div>
-  );
-};
-
-const SavePoemButton = ({ poem, isSaved, onSave, onUnsave, disabled, onSignIn }) => {
-  const handleClick = () => {
-    if (disabled && onSignIn) {
-      onSignIn('Sign in to save your favourites');
-      return;
-    }
-    if (disabled) return;
-
-    if (isSaved) {
-      onUnsave();
-    } else {
-      onSave();
-    }
-  };
-
-  return (
-    <div className="relative flex flex-col items-center gap-1 min-w-[52px]">
-      <button
-        onClick={handleClick}
-        className={`min-w-[46px] min-h-[46px] p-[11px] bg-transparent border-none cursor-pointer transition-all duration-300 flex items-center justify-center rounded-full ${GOLD.goldHoverBg} hover:scale-105`}
-        aria-label={isSaved ? 'Unsave poem' : 'Save poem'}
-      >
-        <Heart
-          size={21}
-          className={`${isSaved ? 'fill-red-500 text-red-500' : GOLD.goldText} transition-all`}
-        />
-      </button>
-      <span
-        className={`font-brand-en text-[8.5px] font-bold tracking-[0.08em] uppercase opacity-60 whitespace-nowrap ${GOLD.goldText}`}
-      >
-        {isSaved ? 'Saved' : 'Save'}
-      </span>
-    </div>
-  );
-};
-
-const DownvoteButton = ({ poem, isDownvoted, onDownvote, onUndownvote, disabled, onSignIn }) => {
-  const handleClick = () => {
-    if (disabled && onSignIn) {
-      onSignIn('Sign in to flag poems');
-      return;
-    }
-    if (disabled) return;
-
-    if (isDownvoted) {
-      onUndownvote();
-    } else {
-      onDownvote();
-    }
-  };
-
-  return (
-    <div className="relative flex flex-col items-center gap-1 min-w-[52px]">
-      <button
-        onClick={handleClick}
-        className={`min-w-[46px] min-h-[46px] p-[11px] bg-transparent border-none cursor-pointer transition-all duration-300 flex items-center justify-center rounded-full ${GOLD.goldHoverBg} hover:scale-105`}
-        aria-label={isDownvoted ? 'Unflag poem' : 'Flag poem'}
-      >
-        <ThumbsDown
-          size={21}
-          className={`${isDownvoted ? 'fill-red-400 text-red-400' : GOLD.goldText} transition-all`}
-        />
-      </button>
-      <span
-        className={`font-brand-en text-[8.5px] font-bold tracking-[0.08em] uppercase opacity-60 whitespace-nowrap ${GOLD.goldText}`}
-      >
-        {isDownvoted ? 'Flagged' : 'Flag'}
-      </span>
-    </div>
-  );
-};
-
-const SavedPoemsView = ({
-  isOpen,
-  onClose,
-  savedPoems,
-  onSelectPoem,
-  onUnsavePoem,
-  theme,
-  currentFontClass,
-}) => {
-  if (!isOpen) return null;
-
-  const formatDate = (dateStr) => {
-    if (!dateStr) return '';
-    const diff = Date.now() - new Date(dateStr).getTime();
-    const mins = Math.floor(diff / 60000);
-    if (mins < 1) return 'Just now';
-    if (mins < 60) return `${mins}m ago`;
-    const hours = Math.floor(mins / 60);
-    if (hours < 24) return `${hours}h ago`;
-    const days = Math.floor(hours / 24);
-    if (days < 30) return `${days}d ago`;
-    return new Date(dateStr).toLocaleDateString();
-  };
-
-  return (
-    <div
-      className="fixed inset-0 z-50 flex items-center justify-center p-4 bg-black/60 backdrop-blur-sm"
-      onClick={(e) => {
-        if (e.target === e.currentTarget) onClose();
-      }}
-      onKeyDown={(e) => {
-        if (e.key === 'Escape') onClose();
-      }}
-    >
-      <div
-        className={`relative w-full max-w-2xl max-h-[85vh] flex flex-col ${theme.glass} ${theme.border} border ${DESIGN.radius} shadow-2xl`}
-      >
-        <div className="flex items-center justify-between p-6 pb-4 border-b border-stone-500/10 flex-shrink-0">
-          <div>
-            <h2 className={`font-amiri text-2xl ${theme.titleColor}`}>قصائدي المحفوظة</h2>
-            <p className={`font-brand-en text-xs ${theme.text} opacity-50 mt-1`}>
-              My Saved Poems ({savedPoems.length})
-            </p>
-          </div>
-          <button
-            onClick={onClose}
-            className="min-w-[44px] min-h-[44px] flex items-center justify-center rounded-full hover:bg-white/10 transition-colors"
-            aria-label="Close"
-          >
-            <X size={20} className={theme.text} />
-          </button>
-        </div>
-
-        <div className="flex-1 overflow-y-auto p-6 custom-scrollbar">
-          {savedPoems.length === 0 ? (
-            <div className="flex flex-col items-center justify-center py-16 gap-4">
-              <Heart size={48} className={`${theme.text} opacity-20`} />
-              <div className="text-center">
-                <p className={`font-amiri text-xl ${theme.text} opacity-40`}>
-                  لا توجد قصائد محفوظة
-                </p>
-                <p className={`font-brand-en text-sm ${theme.text} opacity-30 mt-1`}>
-                  No saved poems yet
-                </p>
-                <p className={`font-brand-en text-xs ${theme.text} opacity-20 mt-3`}>
-                  Tap the heart icon on any poem to save it
-                </p>
-              </div>
-            </div>
-          ) : (
-            <div className="space-y-3">
-              {savedPoems.map((sp) => (
-                <div
-                  key={sp.id}
-                  className={`group ${theme.glass} ${theme.border} border ${DESIGN.radius} p-4 transition-all ${GOLD.goldHoverBorderSubtle}`}
-                >
-                  <button
-                    onClick={() => onSelectPoem(sp)}
-                    className="w-full text-left cursor-pointer bg-transparent border-none p-0"
-                  >
-                    <div className="flex items-start justify-between gap-3">
-                      <div className="flex-1 min-w-0">
-                        <p className={`font-amiri text-sm ${theme.titleColor} font-medium`}>
-                          {sp.poet || 'Unknown'}
-                        </p>
-                        <p className={`font-brand-en text-xs ${theme.text} opacity-50 mt-0.5`}>
-                          {sp.title || ''}
-                        </p>
-                        <p
-                          className={`${currentFontClass} text-sm ${theme.text} opacity-70 mt-2 line-clamp-2`}
-                          dir="rtl"
-                        >
-                          {(sp.poem_text || '').slice(0, 80)}
-                          {(sp.poem_text || '').length > 80 ? '...' : ''}
-                        </p>
-                      </div>
-                    </div>
-                    {sp.saved_at && (
-                      <p className={`font-brand-en text-[10px] ${theme.text} opacity-30 mt-2`}>
-                        {formatDate(sp.saved_at)}
-                      </p>
-                    )}
-                  </button>
-                  <div className="flex justify-end mt-2">
-                    <button
-                      onClick={() => onUnsavePoem(sp)}
-                      className="min-w-[44px] min-h-[44px] flex items-center justify-center rounded-full hover:bg-red-500/10 transition-colors"
-                      aria-label="Remove from saved"
-                    >
-                      <Heart size={16} className="fill-red-500 text-red-500" />
-                    </button>
-                  </div>
-                </div>
-              ))}
-            </div>
-          )}
-        </div>
-      </div>
-    </div>
-  );
-};
-
-/* =============================================================================
-  VERTICAL SIDEBAR (Mobile overflow)
-  =============================================================================
-*/
-
+// Auth components imported from ./components/auth/
 // VerticalSidebar imported from ./components/VerticalSidebar.jsx
 // InsightsDrawer imported from ./components/InsightsDrawer.jsx
 /* =============================================================================
